feat: onboard hyper module
