+++ docs/modules.rst
@@ -8,6 +8,8 @@ Rush Modules
    rush.exess_qmmm
    rush.nnxtb
    rush.hyper
+   rush.hyper_minimize_sumo
+   rush.hyper_run_sumo
    rush.prepare
    rush.auto3d
    rush.mmseqs2
