+++ docs/rush.hyper.rst
@@ -1,5 +1,9 @@
-Hyper
-=====
+:tocdepth: 3
+
+Hyper Solvation
+===============
+
+.. automodule:: rush.hyper._hyper_solvate_sumo
 
 .. currentmodule:: rush.hyper
 
@@ -8,30 +12,10 @@ Run Submission
 
 .. autofunction:: hyper_solvate_sumo
 
-.. autofunction:: hyper_minimize_sumo
-
-.. autofunction:: hyper_run_sumo
-
 Input Types
 -----------
 
 .. autoclass:: HyperConfig
-   :members:
-   :undoc-members:
-
-.. autoclass:: HyperMinimizeConfig
-   :members:
-   :undoc-members:
-
-.. autoclass:: MinimizeInput
-   :members:
-   :undoc-members:
-
-.. autoclass:: HyperRunConfig
-   :members:
-   :undoc-members:
-
-.. autoclass:: RunInput
    :members:
    :undoc-members:
 
@@ -43,21 +27,5 @@ Result Types
    :undoc-members:
 
 .. autoclass:: TRCBatchResultRef
-   :members:
-   :undoc-members:
-
-.. autoclass:: RunOutput
-   :members:
-   :undoc-members:
-
-.. autoclass:: RunOutputRef
-   :members:
-   :undoc-members:
-
-.. autoclass:: RunOutputPaths
-   :members:
-   :undoc-members:
-
-.. autoclass:: RunResultRef
    :members:
    :undoc-members:
