+++ docs/rush.hyper_run_sumo.rst
@@ -1,0 +1,43 @@
+:tocdepth: 3
+
+Hyper Run
+=========
+
+.. automodule:: rush.hyper._hyper_run_sumo
+
+.. currentmodule:: rush.hyper
+
+Run Submission
+--------------
+
+.. autofunction:: hyper_run_sumo
+
+Input Types
+-----------
+
+.. autoclass:: HyperRunConfig
+   :members:
+   :undoc-members:
+
+.. autoclass:: RunInput
+   :members:
+   :undoc-members:
+
+Result Types
+------------
+
+.. autoclass:: RunOutput
+   :members:
+   :undoc-members:
+
+.. autoclass:: RunOutputRef
+   :members:
+   :undoc-members:
+
+.. autoclass:: RunOutputPaths
+   :members:
+   :undoc-members:
+
+.. autoclass:: RunResultRef
+   :members:
+   :undoc-members:
