+++ examples/hyper-minimize_sumo/README.md
@@ -1,0 +1,27 @@
+# Hyper Minimize Sumo Example
+
+Demonstrates `hyper.hyper_minimize_sumo()` with one structure/topology job.
+
+## Quick Start
+
+```bash
+export RUSH_TOKEN="your-token"
+export RUSH_PROJECT="your-project"
+
+python 09_hyper_minimize_sumo.py
+```
+
+## What This Example Covers
+
+1. Building a `MinimizeInput` job from local JSON files
+2. Running Hyper minimization with explicit config bounds
+3. Fetching and saving successful TRC output
+
+## Input Data
+
+- `data/methanol_trc.json`
+- `data/methanol_topology.json`
+
+## Tutorial
+
+See: [Tutorial 9: Hyper Minimize Sumo](../../docs/tutorials/09-hyper-minimize_sumo.md)
