+++ examples/hyper-minimize_sumo/data/methanol_trc.json
@@ -1,0 +1,39 @@
+{
+  "topology": {
+    "schema_version": "0.2.0",
+    "symbols": ["C", "O", "H", "H", "H", "H"],
+    "geometry": [
+      -0.37, 0.01, -0.01,
+      0.92, -0.5, -0.3,
+      -0.52, 0.03, 1.07,
+      -0.47, 1.01, -0.42,
+      -1.12, -0.64, -0.47,
+      1.56, 0.09, 0.12
+    ],
+    "labels": null,
+    "partial_charges": null,
+    "formal_charges": null,
+    "connectivity": null,
+    "stereochemistry": null,
+    "velocities": null,
+    "fragments": null,
+    "fragment_formal_charges": null,
+    "fragment_partial_charges": null,
+    "fragment_multiplicities": null
+  },
+  "residues": {
+    "residues": [[0, 1, 2, 3, 4, 5]],
+    "seqs": ["LIG"],
+    "seq_ns": [1],
+    "insertion_codes": [""],
+    "labeled": null,
+    "labels": null
+  },
+  "chains": {
+    "chains": [[0]],
+    "alpha_helices": null,
+    "beta_sheets": null,
+    "labeled": null,
+    "labels": null
+  }
+}
