+++ examples/hyper-run_sumo/data/methanol_topology.json
@@ -1,0 +1,333 @@
+{
+  "atom_type_count": 4,
+  "charges": [
+    0.28,
+    -0.68,
+    0.0,
+    0.0,
+    0.0,
+    0.4
+  ],
+  "masses": [
+    12.011,
+    15.999,
+    1.008,
+    1.008,
+    1.008,
+    1.008
+  ],
+  "atom_types": [
+    0,
+    1,
+    2,
+    2,
+    2,
+    3
+  ],
+  "bonds": [
+    {
+      "i": 0,
+      "j": 1,
+      "b0": 0.14179999999999998,
+      "kb": 303937.193826
+    },
+    {
+      "i": 0,
+      "j": 2,
+      "b0": 0.1093,
+      "kb": 287014.992228
+    },
+    {
+      "i": 0,
+      "j": 3,
+      "b0": 0.1093,
+      "kb": 287014.992228
+    },
+    {
+      "i": 0,
+      "j": 4,
+      "b0": 0.1093,
+      "kb": 287014.992228
+    },
+    {
+      "i": 1,
+      "j": 5,
+      "b0": 0.0972,
+      "kb": 469365.2642520001
+    }
+  ],
+  "angles": [
+    {
+      "i": 1,
+      "j": 0,
+      "k": 2,
+      "theta0": 1.895026141937883,
+      "ktheta": 470.32508605816537
+    },
+    {
+      "i": 1,
+      "j": 0,
+      "k": 3,
+      "theta0": 1.895026141937883,
+      "ktheta": 470.32508605816537
+    },
+    {
+      "i": 1,
+      "j": 0,
+      "k": 4,
+      "theta0": 1.895026141937883,
+      "ktheta": 470.32508605816537
+    },
+    {
+      "i": 2,
+      "j": 0,
+      "k": 3,
+      "theta0": 1.8995465447005484,
+      "ktheta": 310.7397495595562
+    },
+    {
+      "i": 2,
+      "j": 0,
+      "k": 4,
+      "theta0": 1.8995465447005484,
+      "ktheta": 310.7397495595562
+    },
+    {
+      "i": 3,
+      "j": 0,
+      "k": 4,
+      "theta0": 1.8995465447005484,
+      "ktheta": 310.7397495595562
+    },
+    {
+      "i": 0,
+      "j": 1,
+      "k": 5,
+      "theta0": 1.8588280132515207,
+      "ktheta": 477.55159186187603
+    }
+  ],
+  "proper_dihedrals": [
+    {
+      "i": 2,
+      "j": 0,
+      "k": 1,
+      "l": 5,
+      "phi0": 0.0,
+      "kphi": 1.246832,
+      "mult": 1
+    },
+    {
+      "i": 2,
+      "j": 0,
+      "k": 1,
+      "l": 5,
+      "phi0": 3.141592653589793,
+      "kphi": -0.577392,
+      "mult": 2
+    },
+    {
+      "i": 2,
+      "j": 0,
+      "k": 1,
+      "l": 5,
+      "phi0": 0.0,
+      "kphi": 0.7238319999999999,
+      "mult": 3
+    },
+    {
+      "i": 3,
+      "j": 0,
+      "k": 1,
+      "l": 5,
+      "phi0": 0.0,
+      "kphi": 1.246832,
+      "mult": 1
+    },
+    {
+      "i": 3,
+      "j": 0,
+      "k": 1,
+      "l": 5,
+      "phi0": 3.141592653589793,
+      "kphi": -0.577392,
+      "mult": 2
+    },
+    {
+      "i": 3,
+      "j": 0,
+      "k": 1,
+      "l": 5,
+      "phi0": 0.0,
+      "kphi": 0.7238319999999999,
+      "mult": 3
+    },
+    {
+      "i": 4,
+      "j": 0,
+      "k": 1,
+      "l": 5,
+      "phi0": 0.0,
+      "kphi": 1.246832,
+      "mult": 1
+    },
+    {
+      "i": 4,
+      "j": 0,
+      "k": 1,
+      "l": 5,
+      "phi0": 3.141592653589793,
+      "kphi": -0.577392,
+      "mult": 2
+    },
+    {
+      "i": 4,
+      "j": 0,
+      "k": 1,
+      "l": 5,
+      "phi0": 0.0,
+      "kphi": 0.7238319999999999,
+      "mult": 3
+    }
+  ],
+  "improper_dihedrals": [],
+  "pairs14": [
+    {
+      "i": 2,
+      "j": 5,
+      "c6": 4.274187696047065e-05,
+      "c12": 1.0119320214366275e-08
+    },
+    {
+      "i": 3,
+      "j": 5,
+      "c6": 4.274187696047065e-05,
+      "c12": 1.0119320214366275e-08
+    },
+    {
+      "i": 4,
+      "j": 5,
+      "c6": 4.274187696047065e-05,
+      "c12": 1.0119320214366275e-08
+    }
+  ],
+  "cmap_entries": [],
+  "lj_params": [
+    {
+      "c6": 0.002115016015813491,
+      "c12": 3.94244116886151e-06
+    },
+    {
+      "c6": 0.0016677018383901895,
+      "c12": 2.311249945030495e-06
+    },
+    {
+      "c6": 0.0005432157553294566,
+      "c12": 4.6102104135361286e-07
+    },
+    {
+      "c6": 0.0003954334320525218,
+      "c12": 2.4429962496901167e-07
+    },
+    {
+      "c6": 0.0016677018383901895,
+      "c12": 2.311249945030495e-06
+    },
+    {
+      "c6": 0.00129488850377234,
+      "c12": 1.3138538180583043e-06
+    },
+    {
+      "c6": 0.0004104134700407302,
+      "c12": 2.481369517052484e-07
+    },
+    {
+      "c6": 0.00029313999406590564,
+      "c12": 1.2658970268056165e-07
+    },
+    {
+      "c6": 0.0005432157553294566,
+      "c12": 4.6102104135361286e-07
+    },
+    {
+      "c6": 0.0004104134700407302,
+      "c12": 2.481369517052484e-07
+    },
+    {
+      "c6": 0.00012386870467525756,
+      "c12": 4.249495325323946e-08
+    },
+    {
+      "c6": 8.54837539209413e-05,
+      "c12": 2.023864042873255e-08
+    },
+    {
+      "c6": 0.0003954334320525218,
+      "c12": 2.4429962496901167e-07
+    },
+    {
+      "c6": 0.00029313999406590564,
+      "c12": 1.2658970268056165e-07
+    },
+    {
+      "c6": 8.54837539209413e-05,
+      "c12": 2.023864042873255e-08
+    },
+    {
+      "c6": 5.756897163937681e-05,
+      "c12": 9.17890990287509e-09
+    }
+  ],
+  "cmap_grid_size": 24,
+  "cmap_grids": [],
+  "exclusions": [
+    [
+      0,
+      1,
+      2,
+      3,
+      4,
+      5
+    ],
+    [
+      0,
+      1,
+      2,
+      3,
+      4,
+      5
+    ],
+    [
+      0,
+      1,
+      2,
+      3,
+      4,
+      5
+    ],
+    [
+      0,
+      1,
+      2,
+      3,
+      4,
+      5
+    ],
+    [
+      0,
+      1,
+      2,
+      3,
+      4,
+      5
+    ],
+    [
+      0,
+      1,
+      2,
+      3,
+      4,
+      5
+    ]
+  ]
+}
