+++ examples/hyper-solvate_sumo/README.md
@@ -1,0 +1,26 @@
+# Hyper Solvate Sumo Example
+
+Demonstrates `hyper.hyper_solvate_sumo()` on a small TRC input and validates successful output handling.
+
+## Quick Start
+
+```bash
+export RUSH_TOKEN="your-token"
+export RUSH_PROJECT="your-project"
+
+python 08_hyper_solvate_sumo.py
+```
+
+## What This Example Covers
+
+1. Submitting a Hyper solvation run with explicit `HyperConfig`
+2. Fetching parsed output as `TRC`
+3. Saving the solvated artifact to the Rush workspace
+
+## Input Data
+
+- `data/valid_trc.json`
+
+## Tutorial
+
+See: [Tutorial 8: Hyper Solvate Sumo](../../docs/tutorials/08-hyper-solvate_sumo.md)
